fixed possibly major bug in shape of P1640 spot positions array
the array of spot positions was being reshaped along the wrong axes - the
assumed order was (Ncubes, Nlambda, Nspots, XY) but the actual order
was (Ncubes, Nspots, Nlambda, XY) so when the array was reshaped to (Ncubes x
Nlambda, Nspots, XY) the axis order was wrong.
it's unclear if this will affect the centering and scaling. I don't think it will
because the scaling and centering for each cube are computed before the
wavelength slices are combined. I think this will help when it comes to building
an instrument PSF, though.